pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "emeet-cli"
version = "0.1.0"
description = "CLI to control the eMeet Pixy webcam (zoom, tilt, pan, capture)"
requires-python = ">=3.10"
dependencies = [
    "click>=8.0",
    "pyusb>=1.2",
    "opencv-python>=4.8",
    "pyobjc-framework-AVFoundation>=9.0",
]

[project.scripts]
emeet = "emeet_cli.cli:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["emeet_cli*"]
